+++ src/apps/editor/src/lib/download.ts
@@ -1,5 +1,5 @@
 export default function download(filename: string, data: string | Blob) {
-    if(typeof data !== "string") data = new Blob([data], { type: "text/plain" });
+    if(typeof data === "string") data = new Blob([data], { type: "text/plain" });
   
     const link = document.createElement("a"); // Or maybe get it from the current document
     link.href = URL.createObjectURL(data as Blob);
